+++ packages/br-validators/src/cnpj-motivos.ts
@@ -1,0 +1,1 @@
+export * from './cnpj-motivos/index.js';